Move everything into tools
