+++ indra/tools/machine/default-config.yaml
